% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeInput_Fn.R
\name{MakeInput_Fn}
\alias{MakeInput_Fn}
\title{Build data input for spatial dynamic factor analysis (SDFA)}
\usage{
MakeInput_Fn(Version, Nfactors, DF, loc_xy, method = "mesh",
  Nobsfactors = 0, Kappa_Type = "Constant", ObsModel = NULL,
  Aniso = FALSE, Include_Omega = TRUE, Include_Epsilon = TRUE,
  EncounterFunction = 2, Correlated_Overdispersion = FALSE,
  Include_Phi = TRUE, Include_Rho = TRUE, Use_REML = FALSE, X_ik = NULL,
  X_nl = NULL, X_ntl = NULL, a_n = NULL, YearSet = NULL,
  IndependentTF = c(FALSE, FALSE), CheckForBugs = TRUE,
  CorrGroup_pp = NULL, ...)
}
\arguments{
\item{Version}{a version number (see example for current default).}

\item{Nfactors}{The number of dynamic factors used to approximate spatio-temporal variation}

\item{DF}{a data frame of data where each row is a unique sample and with the following columns
\describe{
  \item{catch}{the observation for each sample}
  \item{year}{a column of years for each sample}
  \item{spp}{a factor specifying the species for that sample}
  \item{sitenum}{a column of years for each sample}
  \item{PredTF}{a vector of 0s or 1st, stating whether a sample is used when fitting the model (PredTF=0) or when evaluating its predictive performance (PredTF=1)}
  \item{TowID}{a vector coding the unit of overdispersion, e.g., tow or vessel (OPTIONAL)}
}}

\item{loc_xy}{Locations for each station}

\item{method}{The method for approximating spatial variation (options: "grid" or "mesh")}

\item{Nobsfactors}{The number of factors used to approximate overdispersion among levels of \code{TowID} (Default=0)}

\item{Kappa_Type}{Whether the decorrelation distance is constant for all factors or not (Default="Constant")}

\item{ObsModel}{The observation model used}

\item{Include_Omega}{Whether to estimate purely spatial variation (Default=TRUE)}

\item{Include_Epsilon, }{Whether to incldue spatio-temporal variation (Default=TRUE)}

\item{EncounterFunction, }{The link between density and encounter probability; 0=two-parameter logistic relationship; 1=two-parameter logistic relationship with less-than-one saturation; 2=one-parameter saturating relationship (Default=1)}

\item{Correlated_Overdispersion, }{Whether to estimate overdispersion (only possible if TowID is present in \code{DF}}

\item{Include_Phi}{Whether to estimate each factor in equilibrium (FALSE) or with a fixed offset from equilibrium (TRUE), Default=TRUE}

\item{Include_Rho}{Whether to estimate the magnitude of temporal autocorrelation (Default=TRUE)}

\item{Use_REML}{Whether to use REML estimation}

\item{X_ik}{A matrix specifying measured variables that affect catchability for each sample (Default is turned off)}

\item{X_nl}{A matrix specifying measured variables that affect density for each sample (Default is an intercept)}

\item{X_ntl}{An array specifying measured variables that affect density and vary over time (Default is off)}

\item{a_n}{A vector giving the area associated with mesh vertex used when calculating indices of abundance (Default is even weighting of sample locations)}

\item{YearSet}{A vector of levels of the \code{year} input that should be modeled (Default is to infer from the year column of DF)}

\item{IndependentTF}{Whether the spatio-temporal variation (IndependentTF[1]) or overdispersion (IndependentTF[2]) is independent among species (Default is both are correlated)}

\item{CheckForBugs}{Whether to check inputs for obvious problems (Default=TRUE)}

\item{CorrGroup_pp}{A matrix filled with integers, for post hoc testing of differences in among-species correlations}
}
\value{
Tagged list containing inputs to function \code{Build_TMB_Fn()}
\describe{
  \item{TmbData}{A tagged list of data inputs}
  \item{TmbParams}{A tagged list with input parameters}
  \item{Random}{A character vector specifying which parameters are treated as random effects}
  \item{Map}{A parameter map, used to turn off or mirror parameter values}
}
}
\description{
\code{MakeInput_Fn} builds a tagged list of inputs for TMB
}

