% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Summary.R
\name{Summary}
\alias{Summary}
\title{Summarize and explain results}
\usage{
Summary(obj, species_names = NULL, rotation_method = "PCA")
}
\arguments{
\item{obj}{the TMB object after it has been previously optimized}

\item{species_names}{an optional character-vector giving species names}

\item{rotation_method}{what rotation method is used when visualizing results (default \code{rotation_method="PCA"})}
}
\value{
Tagged list of useful output
\describe{
  \item{parhat}{a list of fixed-effect estimates, with same format as starting values}
  \item{L_pj}{the loadings matrix, where e.g. \code{L_pj[1,2]} gives the association of species 1 with factor 2}
  \item{Psi}{the array of factors, where e.g. \code{Psi[1,2,3]} gives the value at site 1 of factor 2 in time 3}
  \item{L_pj_rot}{the loadings matrix after rotation, useful for ease of interpretation}
  \item{Psi_rot}{the value of factors, useful for ease of interpretation}
  \item{range_jz}{Decorrelation distance for factor j (1st column: spatial; 2nd column: spatio-temporal}
  \item{theta_npt}{Empirical Bayes prediction of log-density, where e.g., \code{theta_npt[1,2,3]} gives the log-density for site 1 and species 2 in time 3}
  \item{eta_npt}{Empirical Bayes prediction of log-detectability, where e.g., \code{eta_npt[1,2,3]} gives the log-detectability for site 1 and species 2 in time 3}
  \item{logchat_i}{Empirical Bayes prediction of log-expectation for observation i}
  \item{nll_i}{Empirical Bayes prediction of one-half of the deviance residual for observation i}
}
}
\description{
\code{Summary} is intended to help document results, by providing explanation for outputs from \code{obj$report()}
}

